refactor pbt strategies into separate module (ready for reuse) and fix tests after previous annotator refactor
