+++ tests/parsers/test_napoleon_pbt.py
@@ -1,11 +1,10 @@
 import re
 from collections import Counter, OrderedDict
-from functools import partial
 
 import parsy
 import pytest
-from hypothesis import assume, given, note, strategies as st
-from typing import Any, Dict, NamedTuple
+from hypothesis import given, note, strategies as st
+from typing import Dict
 
 from waterloo.parsers.napoleon import (
     args_head,
@@ -23,12 +22,11 @@ from waterloo.parsers.napoleon import (
 from waterloo.types import (
     ArgsSection,
     TypeAtom,
-    TypeDef,
-    SourcePos,
-    VALID_ARGS_SECTION_NAMES,
     VALID_RETURNS_SECTION_NAMES,
 )
-from waterloo.utils import slice_by_pos
+from waterloo.annotator.utils import slice_by_pos
+
+from tests.parsers import strategies
 
 
 """
@@ -40,154 +38,22 @@ They are also reassuringly slow to run ;)
 """
 
 
-class Example(NamedTuple):
-    example: Any
-    context: Dict[str, Any]
-
-
-st_whitespace_char = st.text(' \t', min_size=1, max_size=1)
-
-st_small_lists = partial(st.lists, min_size=0, max_size=3)
-st_small_lists_nonempty = partial(st.lists, min_size=1, max_size=4)
-
-
-@st.composite
-def st_whitespace(draw, min_size=0, max_size=10):
-    """
-    homogenous whitespace of random type (space or tab) and random length
-    (including zero, i.e. '')
-    """
-    n = draw(st.integers(min_value=min_size, max_value=max_size))
-    ws = draw(st_whitespace_char)
-    return ws * n
-
-
-@st.composite
-def st_strip_whitespace(draw, *args, **kwargs):
-    """
-    Drawn from (by default) the full range of Hypothesis' `text` strategy
-    but eliminating initial/trailing whitespace chars, including \n, but
-    not from middle of string.
-    """
-    kwargs['alphabet'] = st.characters(blacklist_characters="\n\r")
-    val = draw(st.text(*args, **kwargs))
-    assume(val.strip() == val)
-    return val
-
-
-st_optional_newline = st.one_of(
-    st.just(''),
-    st.just('\n'),
-)
-
-
-VALID_SECTION_HEAD_TEMPLATE = "{section_name}:{trailing_whitespace}\n"
-
-st_invalid_section_head_template = st.one_of(
-    st.just("{leading_whitespace}{section_name}:{trailing_whitespace}"),
-    st.just("{leading_whitespace}{section_name}{trailing_whitespace}\n"),
-)
-
-st_valid_args_section_name = st.one_of(
-    *(st.just(word) for word in VALID_ARGS_SECTION_NAMES)
-)
-
-st_invalid_args_section_name = st.text().filter(
-    lambda t: t not in VALID_ARGS_SECTION_NAMES
-)
-
-
-@st.composite
-def st_valid_args_head(draw):
-    section_name = draw(st_valid_args_section_name)
-    trailing_whitespace = draw(st_whitespace())
-    return VALID_SECTION_HEAD_TEMPLATE.format(
-        section_name=section_name,
-        trailing_whitespace=trailing_whitespace,
-    )
-
-
-@st.composite
-def st_invalid_args_head_bad_name(draw):
-    section_name = draw(st_invalid_args_section_name)
-    trailing_whitespace = draw(st_whitespace())
-    return VALID_SECTION_HEAD_TEMPLATE.format(
-        section_name=section_name,
-        trailing_whitespace=trailing_whitespace,
-    )
-
-
-@st.composite
-def st_invalid_args_head_bad_template(draw):
-    leading_whitespace = draw(st_whitespace())
-    section_name = draw(st_valid_args_section_name)
-    trailing_whitespace = draw(st_whitespace())
-    template = draw(st_invalid_section_head_template)
-    return template.format(
-        leading_whitespace=leading_whitespace,
-        section_name=section_name,
-        trailing_whitespace=trailing_whitespace,
-    )
-
-
-@given(st_valid_args_head())
+@given(strategies.valid_args_head_f())
 def test_valid_args_head(example):
     result = args_head.parse(example)
     assert result == "Args"  # (Section name has been normalised)
 
 
 @given(st.one_of(
-    st_invalid_args_head_bad_name(),
-    st_invalid_args_head_bad_template(),
+    strategies.invalid_args_head_bad_name_f(),
+    strategies.invalid_args_head_bad_template_f(),
 ))
 def test_invalid_args_head(example):
     with pytest.raises(parsy.ParseError):
         args_head.parse(example)
 
 
-st_valid_returns_section_name = st.one_of(
-    *(st.just(word) for word in VALID_RETURNS_SECTION_NAMES)
-)
-
-st_invalid_returns_section_name = st.text().filter(
-    lambda t: t not in VALID_RETURNS_SECTION_NAMES
-)
-
-
-@st.composite
-def st_valid_returns_head(draw):
-    section_name = draw(st_valid_returns_section_name)
-    trailing_whitespace = draw(st_whitespace())
-    return VALID_SECTION_HEAD_TEMPLATE.format(
-        section_name=section_name,
-        trailing_whitespace=trailing_whitespace,
-    )
-
-
-@st.composite
-def st_invalid_returns_head_bad_name(draw):
-    section_name = draw(st_invalid_returns_section_name)
-    trailing_whitespace = draw(st_whitespace())
-    return VALID_SECTION_HEAD_TEMPLATE.format(
-        section_name=section_name,
-        trailing_whitespace=trailing_whitespace,
-    )
-
-
-@st.composite
-def st_invalid_returns_head_bad_template(draw):
-    leading_whitespace = draw(st_whitespace())
-    section_name = draw(st_valid_returns_section_name)
-    trailing_whitespace = draw(st_whitespace())
-    template = draw(st_invalid_section_head_template)
-    return template.format(
-        leading_whitespace=leading_whitespace,
-        section_name=section_name,
-        trailing_whitespace=trailing_whitespace,
-    )
-
-
-@given(st_valid_returns_head())
+@given(strategies.valid_returns_head_f())
 def test_valid_returns_head(example):
     result = returns_head.parse(example)
     assert result in VALID_RETURNS_SECTION_NAMES
@@ -196,8 +62,8 @@ def test_valid_returns_head(example):
 
 
 @given(st.one_of(
-    st_invalid_returns_head_bad_name(),
-    st_invalid_returns_head_bad_template(),
+    strategies.invalid_returns_head_bad_name_f(),
+    strategies.invalid_returns_head_bad_template_f(),
 ))
 def test_invalid_returns_head(example):
     with pytest.raises(parsy.ParseError):
@@ -206,8 +72,8 @@ def test_invalid_returns_head(example):
 
 @given(
     splat=st.text('*', min_size=0, max_size=4),
-    name=st_strip_whitespace(min_size=0, max_size=10),
-    trailing_ws=st_whitespace(),
+    name=strategies.strip_whitespace_f(min_size=0, max_size=10),
+    trailing_ws=strategies.whitespace_f(),
     newline=st.one_of(st.just(''), st.just('\n')),
 )
 def test_arg_name(splat, name, trailing_ws, newline):
@@ -227,8 +93,8 @@ def test_arg_name(splat, name, trailing_ws, newline):
 
 
 @given(
-    segments=st_small_lists_nonempty(st.text(min_size=0, max_size=8)),
-    trailing_ws=st_whitespace(),
+    segments=strategies.small_lists_nonempty_f(st.text(min_size=0, max_size=8)),
+    trailing_ws=strategies.whitespace_f(),
     newline=st.one_of(st.just(''), st.just('\n')),
 )
 def test_dotted_var_path(segments, trailing_ws, newline):
@@ -251,119 +117,6 @@ def test_dotted_var_path(segments, trailing_ws, newline):
     else:
         with pytest.raises(parsy.ParseError):
             dotted_var_path.parse(example)
-
-
-st_python_identifier = (
-    st
-    .from_regex(r'[^\W0-9][\w]*', fullmatch=True)
-    .filter(lambda s: s.isidentifier())
-)
-
-
-@st.composite
-def st_dotted_var_path(draw):
-    segments = draw(st_small_lists_nonempty(st_python_identifier))
-    return '.'.join(segments)
-
-
-@st.composite
-def st_noargs_typeatom(draw):
-    return TypeAtom(
-        name=draw(st_dotted_var_path()),
-        args=(),
-    )
-
-
-@st.composite
-def st_generic_typeatom(draw, st_children):
-    """
-    A type var with params (i.e. is 'generic'), without being one of the
-    special cases such as homogenous tuple, callable (others?)
-
-    Args:
-        draw: provided by @st.composite
-        st_children: another hypothesis strategy to draw from
-            (first arg to function returned by decorator)
-    """
-    return TypeAtom(
-        name=draw(st_dotted_var_path()),
-        args=draw(st_small_lists(st_children))
-    )
-
-
-@st.composite
-def st_homogenous_tuple_typeatom(draw, st_children):
-    return TypeAtom(
-        name='Tuple',
-        args=(draw(st_children), TypeAtom('...', []))
-    )
-
-
-@st.composite
-def st_callable_typeatom(draw, st_children):
-    args_param = draw(st_small_lists_nonempty(st_children))
-    returns_param = draw(st_children)
-    return TypeAtom(
-        name='Callable',
-        args=(args_param, returns_param)
-    )
-
-
-@st.composite
-def st_type_atom(draw):
-    """
-    Generate arbitrarily nested TypeAtom instances
-    """
-    example = draw(
-        st.recursive(
-            st_noargs_typeatom(),
-            lambda st_children: st.one_of(
-                st_children,
-                st_generic_typeatom(st_children),
-                st_homogenous_tuple_typeatom(st_children),
-                st_callable_typeatom(st_children),
-            ),
-            max_leaves=5,
-        )
-    )
-    return example
-
-
-def _add_arbitrary_whitespace(segment, whitespace, newline):
-    """
-    Adds arbitrary whitespace and optional newlines to a segment
-    from a split TypeAtom string
-    """
-    if segment.startswith('['):
-        return f'{segment}{newline}{whitespace}'
-    elif segment.startswith(','):
-        return f'{segment}{newline}{whitespace}'
-    elif segment.startswith(']'):
-        return f'{newline}{whitespace}{segment}'
-    else:
-        return segment
-
-
-@st.composite
-def st_napoleon_type_annotation(draw):
-    """
-    Generate a type annotation that you might find in a napoleon docstring
-    made from a valid TypeAtom but with arbitrary whitespace in valid locations
-
-    - after a `[`
-    - after a `,`
-    - before a `]`
-    """
-    type_atom = draw(st_type_atom())
-    annotation = type_atom.to_annotation(False)
-    return ''.join(
-        _add_arbitrary_whitespace(
-            segment=segment.strip(),
-            whitespace=draw(st_whitespace()),
-            newline=draw(st_optional_newline),
-        )
-        for segment in re.split(r'([\[\]\,])', annotation)
-    )
 
 
 def _add_normalised_whitespace(segment):
@@ -393,7 +146,7 @@ def assert_annotation_roundtrip(example: str, result: TypeAtom):
     assert normalised == result.to_annotation(False)
 
 
-@given(st_napoleon_type_annotation())
+@given(strategies.napoleon_type_annotation_f())
 def test_type_atom(example):
     """
     Generate an arbitrary ('dirty') type annotation and check that we can
@@ -403,63 +156,7 @@ def test_type_atom(example):
     assert_annotation_roundtrip(example, result)
 
 
-st_rest_of_line = st_strip_whitespace(min_size=0, max_size=80)
-
-
-@st.composite
-def st_arg_description_start(draw):
-    """
-    The rest-of-line that optionally follows an "arg (type)"
-
-    Example:
-        ": blah blah blah"
-    """
-    ws = draw(st_whitespace())
-    trailer = draw(st_rest_of_line)
-    return f":{ws}{trailer}"
-
-
-@st.composite
-def st_arg_name(draw):
-    splat = draw(st.text('*', min_size=0, max_size=2))
-    name = draw(st_python_identifier)
-    return f"{splat}{name}"
-
-
-@st.composite
-def st_annotated_arg(draw):
-    arg_name = draw(st_arg_name())
-    whitespace = draw(st_whitespace(min_size=1))
-    type_annotation = draw(st_napoleon_type_annotation())
-    trailer = draw(
-        st.one_of(
-            st.just(''),
-            st_arg_description_start(),
-        )
-    )
-    type_lines = type_annotation.split("\n")
-    start_pos = SourcePos(
-        row=0,
-        col=len(arg_name) + len(whitespace) + 1
-    )
-    end_pos = SourcePos(
-        row=0 + len(type_lines) - 1,
-        col=len(type_lines[-1])
-    )
-    return Example(
-        f"{arg_name}{whitespace}({type_annotation}){trailer}",
-        {
-            'arg_name': arg_name,
-            'whitespace': whitespace,
-            'type_annotation': type_annotation,
-            'trailer': trailer,
-            'start_pos': start_pos,
-            'end_pos': end_pos,
-        }
-    )
-
-
-@given(annotated_arg_example=st_annotated_arg())
+@given(annotated_arg_example=strategies.annotated_arg_f())
 def test_arg_type_annotated(annotated_arg_example):
     """
     Test that we can parse the first line of an arg and its description
@@ -476,10 +173,10 @@ def test_arg_type_annotated(annotated_arg_example):
 
 
 @given(
-    arg_name=st_arg_name(),
+    arg_name=strategies.arg_name_f(),
     trailer=st.one_of(
         st.just(''),
-        st_arg_description_start(),
+        strategies.arg_description_start_f(),
     )
 )
 def test_arg_type_no_annotation(arg_name, trailer):
@@ -493,27 +190,9 @@ def test_arg_type_no_annotation(arg_name, trailer):
     assert result['arg'] == arg_name
 
 
-@st.composite
-def st_rest_of_line_with_insertion(draw):
-    """
-    Insert something that looks like an args head in the middle of a line
-    which should be ignored
-    """
-    line = draw(st_strip_whitespace(min_size=1, max_size=80))
-    insertion_index = draw(st.integers(min_value=0, max_value=len(line)))
-    head_name = draw(st_valid_args_section_name)
-    return f"{line[:insertion_index]}{head_name}: {line[insertion_index:]}"
-
-
-st_ignored_line = st.one_of(
-    st_rest_of_line,
-    st_rest_of_line_with_insertion(),
-)
-
-
 @given(
-    indent=st_whitespace(),
-    line_to_ignore=st_ignored_line,
+    indent=strategies.whitespace_f(),
+    line_to_ignore=strategies.ignored_line,
 )
 def test_ignored_line(indent, line_to_ignore):
     """
@@ -527,9 +206,9 @@ def test_ignored_line(indent, line_to_ignore):
 
 
 @given(
-    indent=st_whitespace(),
+    indent=strategies.whitespace_f(),
     blank_lines=st.text('\n', min_size=0, max_size=2),
-    ignored_lines=st_small_lists(st_ignored_line),
+    ignored_lines=strategies.small_lists_f(strategies.ignored_line),
 )
 def test_ignored_lines(
     indent,
@@ -547,64 +226,6 @@ def test_ignored_lines(
     result, remainder = parser.parse_partial(example)
     assert result == [""] * tbc_line_count
     assert remainder == unconsumed
-
-
-@st.composite
-def st_annotated_arg_full(draw, initial_indent, indent, row_offset=0):
-    first_line, arg_context = draw(st_annotated_arg())
-    wrapped_lines = draw(st_small_lists(st_rest_of_line))
-
-    offset = SourcePos(row_offset, len(initial_indent) + len(indent))
-    arg_context['start_pos'] += offset
-    arg_context['end_pos'] += offset
-
-    if wrapped_lines:
-        continuation = "\n".join(
-            f"{initial_indent}{indent*2}{indent}{line}"
-            for line in wrapped_lines
-        )
-        return Example(
-            f"{initial_indent}{indent}{first_line}\n{continuation}\n",
-            arg_context
-        )
-    else:
-        return Example(
-            f"{initial_indent}{indent}{first_line}\n",
-            arg_context
-        )
-
-
-@st.composite
-def st_args_section(draw, initial_indent=None, row_offset=0):
-    """
-    Header plus a list of annotated args
-    """
-    if initial_indent is None:
-        initial_indent = draw(st_whitespace())
-    if initial_indent:
-        indent = initial_indent
-    else:
-        indent = draw(st_whitespace(min_size=2))
-
-    args_head = draw(st_valid_args_head())  # includes line-break
-    annotated_args = draw(
-        st_small_lists_nonempty(
-            st_annotated_arg_full(
-                initial_indent=initial_indent,
-                indent=indent,
-                row_offset=row_offset + 1,
-            ),
-            unique_by=lambda a: a.context['arg_name']
-        )
-    )
-    args_str = "\n".join(arg[0] for arg in annotated_args)
-    return Example(
-        f"{initial_indent}{args_head}{args_str}",
-        {
-            'args_head': args_head,
-            'annotated_args': annotated_args,
-        }
-    )
 
 
 def _validate_args_section(example, result, context):
@@ -625,95 +246,11 @@ def _validate_args_section(example, result, context):
         assert slice_by_pos(example, start, end) == type_annotation_str
 
 
-@given(st_args_section())
+@given(strategies.args_section_f())
 def test_p_arg_list(args_section):
     example, context = args_section
     result = p_arg_list.parse(example)
     _validate_args_section(example, result, context)
-
-
-@st.composite
-def st_annotated_return(draw):
-    type_annotation = draw(st_napoleon_type_annotation())
-    trailer = draw(
-        st.one_of(
-            st.just(''),
-            st_arg_description_start(),
-        )
-    )
-    type_lines = type_annotation.split("\n")
-    start_pos = SourcePos(
-        row=0,
-        col=0
-    )
-    end_pos = SourcePos(
-        row=0 + len(type_lines) - 1,
-        col=len(type_lines[-1])
-    )
-    return Example(
-        f"{type_annotation}{trailer}",
-        {
-            'type_annotation': type_annotation,
-            'trailer': trailer,
-            'start_pos': start_pos,
-            'end_pos': end_pos,
-        }
-    )
-
-
-@st.composite
-def st_annotated_return_full(draw, initial_indent, indent, row_offset=0):
-    first_line, context = draw(st_annotated_return())
-    wrapped_lines = draw(st_small_lists(st_rest_of_line))
-
-    offset = SourcePos(row_offset, len(initial_indent) + len(indent))
-    context['start_pos'] += offset
-    context['end_pos'] += offset
-
-    if wrapped_lines:
-        continuation = "\n".join(
-            f"{initial_indent}{indent*2}{indent}{line}"
-            for line in wrapped_lines
-        )
-        return Example(
-            f"{initial_indent}{indent}{first_line}\n{continuation}\n",
-            context
-        )
-    else:
-        return Example(
-            f"{initial_indent}{indent}{first_line}\n",
-            context
-        )
-
-
-@st.composite
-def st_returns_section(draw, initial_indent=None, row_offset=0):
-    """
-    Header plus the return type with optional description
-    """
-    if initial_indent is None:
-        initial_indent = draw(st_whitespace())
-    if initial_indent:
-        indent = initial_indent
-    else:
-        indent = draw(st_whitespace(min_size=2))
-
-    returns_head = draw(st_valid_returns_head())
-    annotated_return = draw(
-        st_annotated_return_full(
-            initial_indent=initial_indent,
-            indent=indent,
-            row_offset=row_offset + 1,
-        )
-    )
-
-    return Example(
-        f"{initial_indent}{returns_head}{annotated_return[0]}",
-        {
-            'returns_head': returns_head,
-            'annotated_return': annotated_return,
-        }
-    )
 
 
 def _validate_returns_section(example, result, context):
@@ -727,71 +264,14 @@ def _validate_returns_section(example, result, context):
     assert slice_by_pos(example, start, end) == type_annotation_str
 
 
-@given(st_returns_section())
+@given(strategies.returns_section_f())
 def test_p_returns_block(returns_section):
     example, context = returns_section
     result = p_returns_block.parse(example)
     _validate_returns_section(example, result, context)
 
 
-@st.composite
-def st_napoleon_docstring(draw):
-    initial_indent = draw(st_whitespace())
-
-    _intro_lines = draw(st_small_lists(st_ignored_line))
-    intro = "\n".join(_intro_lines)
-    # last ignored_line has no trailing \n
-    gap_1 = draw(st.text('\n', min_size=1, max_size=3)) if intro else ''
-
-    row_offset = len(_intro_lines) + len(gap_1)
-    args_section = draw(
-        st.one_of(
-            st_args_section(initial_indent, row_offset=row_offset),
-            st.just(Example('', {}))
-        )
-    )
-    gap_2 = (
-        draw(st.text('\n', min_size=0, max_size=2))
-        if args_section[0]
-        else ''
-    )
-
-    row_offset += len(args_section[0].split("\n")) + len(gap_2)
-    returns_section = draw(
-        st.one_of(
-            st_returns_section(initial_indent, row_offset=row_offset),
-            st.just(Example('', {}))
-        )
-    )
-    gap_3 = (
-        draw(st.text('\n', min_size=0, max_size=2))
-        if returns_section[0]
-        else ''
-    )
-
-    following = "\n".join(
-        draw(st_small_lists(st_ignored_line))
-    )
-
-    example = (
-        f"{intro}"
-        f"{gap_1}"
-        f"{args_section[0]}"
-        f"{gap_2}"
-        f"{returns_section[0]}"
-        f"{gap_3}"
-        f"{following}"
-    )
-    return Example(
-        example=example,
-        context={
-            'args_section': args_section,
-            'returns_section': returns_section,
-        }
-    )
-
-
-@given(st_napoleon_docstring())
+@given(strategies.napoleon_docstring_f())
 def test_docstring_parser(docstring):
     example, context = docstring
 
@@ -801,7 +281,7 @@ def test_docstring_parser(docstring):
         note(f"args_section: {context['args_section'][0]}")
         _validate_args_section(
             example,
-            result.args,
+            result.arg_types,
             context['args_section'].context,
         )
 
@@ -809,6 +289,6 @@ def test_docstring_parser(docstring):
         note(f"returns_section: {context['returns_section'][0]}")
         _validate_returns_section(
             example,
-            result.returns,
+            result.return_type,
             context['returns_section'].context,
         )
